chore(server): move reset statut fiab
